+++ Quaver/Core/Router.php
@@ -1,9 +1,30 @@
 <?php
-
-/*
- * Copyright (c) 2014 Alberto González
- * Distributed under MIT License
- * (see README for details)
+/**
+ * Quaver Framework
+ *
+ * @author      Alberto González <quaver@millolab.com>
+ * @copyright   2014 Alberto González
+ *
+ * MIT LICENSE
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
 namespace Quaver\Core;
@@ -14,11 +35,11 @@ use Symfony\Component\Yaml\Exception\ParseException;
 use Quaver\App\Model\User;
 
 /**
- * Router class.
+ * Router class
+ * @package Quaver
  */
 class Router
 {
-    public $version = '0.10';
     public $routes = null;
     public $modules = null;
 
@@ -350,7 +371,7 @@ class Router
                             $controllerLoader->setView($controllerData['controllerView']);
                         }
 
-                        $controllerLoader->$controllerData['actionName']();
+                        return $controllerLoader->$controllerData['actionName']();
                     }
                 }
             } else {
@@ -362,7 +383,7 @@ class Router
                         $controllerLoader->setView($controllerData['controllerView']);
                     }
 
-                    $controllerLoader->$controllerData['actionName']();
+                    return $controllerLoader->$controllerData['actionName']();
                 }
             }
         } catch (\Quaver\Core\Exception $e) {