+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [0.11] - 2015-10-15
+### Changed
+ - Bootstrap model change to App
+ - Now App runs Quaver instances
+
+
 ## [0.10] - 2015-10-14
 ### Added
  - New Model method: getArrayList